# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ZS MT5 Bridge API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ea/reports:
    get:
      operationId: listEaReports
      tags: [ea]
      summary: Get recent EA event reports
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: EA report list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EaReportListResponse"
...
components:
  schemas:
...
    EaReport:
      type: object
      properties:
        id:
          type: integer
        eventType:
          type: string
        symbol:
          type: string
        direction:
          type: ["string", "null"]
        setup:
          type: ["string", "null"]
        score:
          type: ["integer", "null"]
        entry:
          type: ["string", "null"]
        sl:
          type: ["string", "null"]
        tp1:
          type: ["string", "null"]
        tp2:
          type: ["string", "null"]
        tp3:
          type: ["string", "null"]
        closePrice:
          type: ["string", "null"]
        plDollars:
          type: ["string", "null"]
        closeReason:
          type: ["string", "null"]
        holdMinutes:
          type: ["integer", "null"]
        tpLevel:
          type: ["string", "null"]
        rsi:
          type: ["string", "null"]
        adx:
          type: ["string", "null"]
        atr:
          type: ["string", "null"]
        buyScore:
          type: ["integer", "null"]
        sellScore:
          type: ["integer", "null"]
        bullCount:
          type: ["integer", "null"]
        bearCount:
          type: ["integer", "null"]
        srStatus:
          type: ["string", "null"]
        mt5Ticket:
          type: ["integer", "null"]
        sessionOk:
          type: ["boolean", "null"]
        totalSignals:
          type: ["integer", "null"]
        winCount:
          type: ["integer", "null"]
        lossCount:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - eventType
        - symbol
        - createdAt
    EaReportListResponse:
      type: object
      properties:
        ok:
          type: boolean
        reports:
          type: array
          items:
            $ref: "#/components/schemas/EaReport"
      required:
        - ok
        - reports
...