# lib/api-zod/src/generated/types/eaReport.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * ZS MT5 Bridge API
 * OpenAPI spec version: 0.1.0
 */

export interface EaReport {
  id: number;
  eventType: string;
  symbol: string;
  /** @nullable */
  direction?: string | null;
  /** @nullable */
  setup?: string | null;
  /** @nullable */
  score?: number | null;
  /** @nullable */
  entry?: string | null;
  /** @nullable */
  sl?: string | null;
  /** @nullable */
  tp1?: string | null;
  /** @nullable */
  tp2?: string | null;
  /** @nullable */
  tp3?: string | null;
  /** @nullable */
  closePrice?: string | null;
  /** @nullable */
  plDollars?: string | null;
  /** @nullable */
  closeReason?: string | null;
  /** @nullable */
  holdMinutes?: number | null;
  /** @nullable */
  tpLevel?: string | null;
  /** @nullable */
  rsi?: string | null;
  /** @nullable */
  adx?: string | null;
  /** @nullable */
  atr?: string | null;
  /** @nullable */
  buyScore?: number | null;
  /** @nullable */
  sellScore?: number | null;
  /** @nullable */
  bullCount?: number | null;
  /** @nullable */
  bearCount?: number | null;
  /** @nullable */
  srStatus?: string | null;
  /** @nullable */
  mt5Ticket?: number | null;
  /** @nullable */
  sessionOk?: boolean | null;
  /** @nullable */
  totalSignals?: number | null;
  /** @nullable */
  winCount?: number | null;
  /** @nullable */
  lossCount?: number | null;
  createdAt: Date;
}