# lib/api-zod/src/generated/types/eaReportListResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * ZS MT5 Bridge API
 * OpenAPI spec version: 0.1.0
 */
import type { EaReport } from './eaReport';

export interface EaReportListResponse {
  ok: boolean;
  reports: EaReport[];
}